초기 설정 코드 위치 변경
Form1.cs에서 WhereAmI.cs로 변경
